+++ sphinxcontrib/sphinx_bazel/domain/object.py
@@ -30,10 +30,18 @@ class BazelObject(ObjectDescription):
     Description of a general bazel object
     """
     option_spec = {
-        'workspace': directives.flag,
-        'workspace_path': directives.flag,
-        'implementation': directives.unchanged,  # Used by bazel:rule to print implementation function
+        'path': directives.unchanged,  # Can be used to specify local, not-valid workspace
+        'implementation': directives.unchanged,  # Used by bazel:rule to define implementation function
+        'show_workspace': directives.flag,
+        'show_workspace_path': directives.flag,
+        'show_implementation': directives.flag,
     }
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        self.implementation = None
+        self.specific_workspace_path = None
 
     def get_signature_prefix(self, sig):
         # type: (str) -> str
@@ -52,7 +60,7 @@ class BazelObject(ObjectDescription):
     def handle_signature(self, sig, signode):
         """
         Transform a Bazel signatur into RST nodes.
-        
+
         :param sig:
                 :param signode:
         :return:
@@ -63,44 +71,58 @@ class BazelObject(ObjectDescription):
             raise ValueError
         package, after_package, target, after_target, internal = m.groups()
 
-        try:
-            signode['workspace'] = self.env.ref_context['bazel:workspace']
-        except KeyError:
-            logger.error("No workspace defined before given Bazel element on current page.")
-        
+        # Let's see if we have to use a specific workspace path or if we have to use the latest defined workspace
+        self.specific_workspace_path = self.options.get('path', None)
+        if self.specific_workspace_path:
+            signode['workspace'] = self.specific_workspace_path
+        else:
+            try:
+                signode['workspace'] = self.env.ref_context['bazel:workspace']
+            except KeyError:
+                logger.error("No workspace defined before given {} on page {} line {}".format(
+                    self.name, self.state.document.current_source, self.state.document.current_line))
+
         signode['package'] = package
         signode['target'] = target
-        
+
         sig_text = '//{}'.format(package)
         if target:
             sig_text += ':{}'.format(target)
         if internal:
             sig_text += ':{}'.format(internal)
         signode += addnodes.desc_name(sig_text, sig_text)
-        
-        if self.options.get('workspace', False) is None:  # if flag is set, value is None
+
+        if self.options.get('show_workspace', False) is None:  # if flag is set, value is None
             ws_string = 'workspace: {}'.format(signode['workspace'])
             self._add_signature_detail(signode, ws_string)
-            
-        if self.options.get('workspace_path', False) is None:  # if flag is set, value is None
-            current_ws = self.env.ref_context['bazel:workspace']
-            ws_obj = self.env.domaindata['bazel']['workspaces'][current_ws]
-            ws_path = ws_obj[1]  # See workspace.py for details about stored data
+
+        if self.options.get('show_workspace_path', False) is None:  # if flag is set, value is None
+            # If no extra workspace path was defined via :path:
+            if self.specific_workspace_path is None:
+                # Get the path of the current/latest workspace
+                current_ws = self.env.ref_context['bazel:workspace']
+                ws_obj = self.env.domaindata['bazel']['workspaces'][current_ws]
+                ws_path = ws_obj[1]  # See workspace.py for details about stored data
+            else:
+                ws_path = self.specific_workspace_path
             ws_path_string = 'workspace path: {}'.format(ws_path)
             self._add_signature_detail(signode, ws_path_string)
 
         rule_impl = self.options.get('implementation', "")
         if rule_impl:  # if flag is set, value is None
-            impl_string = 'implementation: {}'.format(rule_impl)
+            self.implementation = rule_impl
+
+        if self.options.get('show_implementation', False) is None:
+            impl_string = 'implementation: {}'.format(self.implementation)
             self._add_signature_detail(signode, impl_string)
-        
+
         return sig, sig
 
     def _add_signature_detail(self, signode, text):
         """
         Create a additional line under signature.
         Used to show additional details of an object like workspace name or path
-        
+
         :param signode: node to add the line add the ending
         :param text: Text inside the new line
         :return: None